Move things to shared folder
