+++ front/app/containers/Admin/projects/_shared/components/AnalysisBanner/index.tsx
