+++ front/app/containers/Admin/projects/_shared/components/AnalysisBanner/utils.ts
