+++ front/app/containers/Admin/projects/_shared/components/PDFExportModal/MultilocFieldCollapsible.tsx
