+++ front/app/containers/Admin/projects/_shared/components/ProjectRow/GroupsTag.tsx
