+++ front/app/containers/Admin/projects/_shared/components/ProjectRow/ProjectRow.test.tsx
