+++ front/app/containers/Admin/projects/_shared/components/PublicationStatusLabel.tsx
