+++ front/app/containers/Admin/projects/_shared/types.ts
@@ -1,0 +1,8 @@
+import { IUpdatedProjectProperties } from 'api/projects/types';
+
+import { ISubmitState } from 'components/admin/SubmitWrapper';
+
+export type TOnProjectAttributesDiffChangeFunction = (
+  projectAttributesDiff: IUpdatedProjectProperties,
+  submitState?: ISubmitState
+) => void;
